// src/types/Person.tsx
export type Person = {
    name: string;
    linkedInUrl: string;
    githubUrl: string;
    upworkUrl: string;
    description: string;
    location: string;
    email:string;
    phone:string;
    experience:string[],
    education:string[],
    mobDescription: string,
    mobDetail: string,
}